refactor serializing and improve customizability.
By refactoring the serializing mechanism and using adapters we provide
better dependency injection entry points and have a clearer component
design.

** IFieldSerializer **
The field serializer adapter adapts the dexterity field, the context and
the request. It's duty is to convert the value of the adapted field on
the current context to a data structure which is supported by the JSON
module.
This provides the entry point for customizing how a certain field or a
type of fields should be represented in the JSON.

** IJsonCompatible **
The json compatibel adapter converts values. It does not know anything
about fields and has no context. The idea here is that is possible to
recursively convert data. This is especially important for supporting
list fields, object fields, data grid fields etc.

Specification: https://github.com/plone/plone.restapi/issues/38