Fix Plone 5 test by using assertDictContainsSubset
instead of assertDictEquals.

We use assertDictContainsSubset in order not to have the
Plone 5 tests fail because the automatic ID generation for the
stock 'Document' type works differently in Plone 5 than Plone 4.

A better solution for this (to get stable test behavior across Plone
versions) is to create our own FTIs for tests instead of using the
stock types from Plone (which may be named the same, but behave
differently).