Fail tests on uncommitted changes to docs/source/_json:
This adds a CI test that will fail if there are uncommitted changes
(dirty working dir) to any of the files in docs/source/_json
after running the tests from test_documentation.py.

The purpose of this is to stop having the dynamically produced
request/response dumps that are included in the docs get out of
sync with the actual implementation.