Batching: Preserve list-like query string params when canonicalizing URLs:
Query string parameters can be repeated with the same key, but different
values, which is a de-facto standard for passing lists of values in
query strings.

These got mangled when creating canonical batch URLs (and other URLs
derived from the canonical batch URL, like 'next' and 'first' links)
because we turned the result of `parse_qsl` into a dict.

This change makes sure these list-like query string parameters are
preserved.