Fixed tests after German translation update.
This commit in plone.app.locales changes several German strings:
https://github.com/collective/plone.app.locales/commit/426adfc06bbc94eeddf350ba6d7c93059372c6fd
Reviewer was mostly called 'Redacteur', and is called simply 'Reviewer' now.
And in one case it had been translated as 'Veröffentlichen', which is what plone.restapi was testing,
but that one means 'make public', so it wasn't a good translation.

Updated the test to check for either of the two translations instead,
because the tests need to pass in other Plone versions as well.