Tests: Freeze creation and modification dates in a more targeted way:
Instead of using a library like `freezegun` to freeze time
entirely during tests, be more selective in our patching.

Specifically monkey-patching a few accessors (getter methods) in a
targeted way like this was chosen as an alternative to freezing time
entirely using libraries like `freezegun`, because that approach has
caused issues in the past with code that assumes a monotonic clock, like
the generation of ZODB transaction IDs.