Remove IAPIRequest interface
It's not used in plone.restapi and should be imported from plone.rest.interfaces instead. If anybody was using this marker, it didn't do anything.

We coul alias/import it from plone.rest to here, but then it suddenly would start to work..... I don't know if that's better than removing it.