Patch Plone 4.3's _install_profile_info method to list the default profile of plone.restapi in the first position.
Plone 4.3 considers that the first profile, after ordering all profiles
in alphabetical order, is the profile to be installed when installing
the product. Since plone.restapi has the profile blocks, it is
installed, instead of the default profile. This patch puts the default
profile in the first position, so that it is the installed profile.