Fix `@history` when full history is empty.
The re-used viewlet returns `None` in its `fullHistory` when there is
no (visible) workflow history and when there is no (visible) revision
history.

We did not take this into account and always attempt to iterate the
returned value, which will fail for `None` for obvious reasons.

I'm adding a simple fix for the issue by initializing the value to an
empty list in such cases.