Fix intermittent failures in test_batching
The default ordering of Plone is by relevance. Therefore, it is not
guaranteed that all searches without sort_on, have the same order.

Listings of contents of folders and collections with batch don't accept
sort_on. In such cases, without sort_on it is not possible to force a
different order of relevance.

Because of this, this test had intermittent failures.

This fixes jobs like:

https://github.com/plone/plone.restapi/runs/2577842538#step:11:65