Use checkout of plone.restapi for testing 5.2.
This was already done correctly for the other Plone versions.
Done for all versions in the same way now:
first extend the dist.plone.org versions and then extend base.cfg, which unpins plone.restapi.

Also: test with latest Plone 5.2.5.
And remove pins that are already in 5.2.5 with a good version.