Fix random failures on test_services_types
plone.dexterity 2.10+ now uses the _p_mtime variable in the schema
names. As this variable is related to date/time and the tests are
executed very fast, some schema had the same name, when they shouldn't
be. This was causing problems in the tests. So we introduce sleeps in
Plone 5.2 with Python 2, so that the schemas don't have the same names.
In Python 3 we don't have this problem because the precision of _p_mtime
is higher in Python 3.