Fix random failure in test_querystringsearch_complex
When we don't have the sort_on parameter, it's not guaranteed that the
search result will always come in the same order. This was causing the
test to fail randomly. So we add the sort_on parameter to ensure that
the result always comes in the same order.

This fixes the error:

AssertionError: 'Test Document 8' != 'Test Document 4'