Fix random failures in tests (#1270)
* Fix random failure in test_serialize_to_json_collection_include_items

When we don't have the sort_on parameter, it's not guaranteed that the
search result will always come in the same order. This was causing the
test to fail randomly. So we sort the titles before comparing the
result.

This fixes the error:

AssertionError: 'Document 2' != 'Document 1'

* Fix random failure in test_querystringsearch_complex

When we don't have the sort_on parameter, it's not guaranteed that the
search result will always come in the same order. This was causing the
test to fail randomly. So we add the sort_on parameter to ensure that
the result always comes in the same order.

This fixes the error:

AssertionError: 'Test Document 8' != 'Test Document 4'