fix(auth): Set both classic and API login cookies
Use the appropriate PAS interface for setting the cookies used by both classic Plone and
Volto's React component.  Move the API token generation out of the login API endpoint
view and into the PAS plugin where it belongs.  Then it will be per the PAS plugin
configuration to set the cookie.  The plugin then also makes the token available to the
login API endpoint view via the request so it can then also return the token in the JSON
response.