test(deprecation): Fix some warnings from our code
Resolve all the deprecation warnings that originate in this package's code that are
exposed by running the tests that do not stem from backwards compatibility we still
support.

IOW, warnings are still emitted that stem from code that needs to work with older
versions we still support and cannot be updated without breaking that backwards
compatibility.  It seems better to me to leave the warnings in place than litter our
code with BBB conditionals.