fix(log): Log error conditions in the server logs
An error that indicates incorrect configuration, set up, or otherwise something an
administrator would probably want to know, should also be logged to the server logs.
Those log messages should including enough information for an administrator to find the
specific instance of the issue (e.g. which Plone portal in a multi-portal ZODB).