build(versions): Cleanup unused buildout config
Chanced across this while updating the `Products.PluggableAuthService` version.  The
`./versions.cfg` file isn't referenced anywhere that I can see.  A `./versions.cfg` file
is a convention in the Plone world so it's misleading to have an unused one lying
around.  A developer might reasonable assume that a change in that file should take
effect.