Plone 6 buildout: extend https://dist.plone.org/release/6.0-dev/versions.cfg
And include this in the find-links.
Yesterday I updated the versions, and it includes an internal release of Products.CMFPlone, which is only uploaded there.
This fixes breakage caused today by merging the plone.base PLIP 3395:
our auto-checkout of Products.CMFPlone depends on plone.base but there is no release of it yet.

Alternative fix would be to pin our CMFPlone checkout to commit 1103e16165ede181eec0b1a27b9c5a64e0046172,
which is basically the same code as from the internal release.