Split CI of Plone 5 and 6 in two files.
We need a newer zc.buildout version for Plone 6, otherwise you get an error parsing Zope versions:

    NameError: name 'python37' is not defined

I introduced `requirement-60.txt` for this.
To keep the CI tests in one yaml file, I would have needed to rename `requirements.txt` to `requirements-52.txt`,
but that would keep it trickier to keep syncing various files in this repo with a central kitconcept repo, if we still want this.
So I split the files in two.

Also fixed the Makefile for Plone 6.
Until now it was first running `bin/buildout` before running `bin/buildout -c plone-6.0.x.cfg`, and this easily failed.