Tests: patch unique url for scale in old or new way.
This is only in serializer tests for images.

uuid4 is only used until plone.scale 6.0.0a4.
https://github.com/plone/plone.scale/pull/57 introduces a new way.

We also patch the _modified_since method to always return True.
Otherwise you may get info from a different scale back, precisely because we give all scales the same "unique" id,
which then is of course no longer unique, making the logic unstable.
This is needed for the newer plone.scale versions, but should be perfectly fine for the older ones.