Fix pinns (#1516)
* Uses Plone 6 in requirements.txt

* Use specific version of Plone in requirements.txt

As we are using specific version in versions.cfg, we need to use
specific versions in requirements.txt as well, to avoid problems in the
future.

* Remove Plone 3.7 specific pinns in Plone 6

Package is no longer tested in Plone 3.7 in Plone 6

* Remove pinn from cffi in Plone 6

Zope 5.6 already has a version of this package. See:

https://github.com/zopefoundation/Zope/blob/5.6/versions.cfg#L16

* Removes specific Python 3.6 pinns in Plone 5.2

Python 3.6 is no longer tested in Plone 5.2

* Remove pinn from cffi in Plone 5.2

Zope 4.8.2 already has a version of this package. See:

https://github.com/zopefoundation/Zope/blob/4.8.2/versions.cfg#L17

* Add changelog